🗃  Modified: root/.local/bin/root_clean.sh root/.l... 🗃
 Modified: root/.local/bin/root_clean.sh
 Modified: root/.local/bin/run-os-update

+++ root/.local/bin/root_clean.sh
@@ -103,15 +103,13 @@ else
     rm -Rf $PIDFILE
     rm -Rf $BASEDIR/run/*
     poweroff -h
-    exit $?
 
-  else
+  elif [ -z "$no_reboot" ] || [[ "$1" = *no-reboot ]]; then
     echo "restarting system in 10 seconds"
     sleep 10
     rm -Rf $PIDFILE
     rm -Rf $BASEDIR/run/*
     reboot
-    exit $?
   fi
 fi
 exit $?
